Cleanup cloud command graph & add player parser
